refactor(experiments): extract ExperimentRolloutCard, address review feedback
- Extract rollout card into ExperimentRolloutCard component
- Use isValidPercentage helper in CreateExperimentWizard
- Replace inline styles with w-25 utility class
- Fix allocation display to read from experiment_rollout when present
- Move SRM warning outside scorecard table with reduced font size
